Add back support for Ember 3.16
As discussed in https://github.com/josemarluedke/ember-focus-trap/pull/33#issuecomment-847974000, this introduces a conditional to support Ember 3.16+ while still using the latest modifier capabilities when possible. `@embroider/macros` was chosen for this instead of the usual `ember-compatibility-helpers`, as that seems to more future-proof solution to me, especially allowing an easier transition to an eventual native "v2" addon.